Add email column to masked_people DB view
Note I had to remove the DEFINER statement from the structure.sql.

```
/*!50001 CREATE ALGORITHM=UNDEFINED */
/*!50013 DEFINER=`sharetribe`@`%` SQL SECURITY DEFINER */
/*!50001 VIEW `masked_people` AS select `people`.`id` AS `id
```

I did that when we created but with every update Scenic adds it and this
breaks in production. We definitely need to find a better way. Either
ensure the same DEFINER user is present in production or make Scenic
skip that statement (I'm not sure it can).