Don't specify RAILS_ENV in containers
When you run the specs in a container some gems fail to get loaded due
to the fact that RAILS_ENV is development instead of test. Most of the
times you won't think that's the root cause and have a hard time.

It's better to leave it empty so that Rails assumes development mode and
RSpec sets test mode when executed.